+++ pom.xml
@@ -15,7 +15,7 @@
 	<name>examserver</name>
 	<description>backend code for exam portal</description>
 	<properties>
-		<java.version>14</java.version>
+		<java.version>11</java.version>
 	</properties>
 	<dependencies>
 		<dependency>
